requirements.txt
asyncio
pytest==7.4.3
pytest-asyncio==0.21.1
pydantic==2.5.2
python-json-logger==2.0.7
openai>=1.35.0
aiohttp==3.9.3
httpx>=0.25.0
python-dotenv==1.0.0